requirements.txt
asgiref==3.7.2
astroid==2.15.6
certifi==2023.7.22
dill==0.3.7
Django==4.2.3
django-elasticsearch-dsl==7.3
django-rainbowtests==0.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
password-validator==1.0
platformdirs==3.10.0
PyJWT==2.8.0
pylint==2.17.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.7.1
urllib3==1.26.16
wrapt==1.15.0
